backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
itsdangerous
django-environ
django-autoslug